// src/tmdb/schema.d.ts
export interface operations {
    "search-movie": {
        parameters: {
            query: {
                query: string;
                include_adult?: boolean;
                language?: string;
                primary_release_year?: string;
                page?: number;
                region?: string;
                year?: string;
            };
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description 200 */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": {
                        /**
                         * @default 0
                         * @example 1
                         */
                        page: number;
                        results?: {
                            /**
                             * @default true
                             * @example false
                             */
                            adult: boolean;
                            /** @example /hZkgoQYus5vegHoetLkCJzb17zJ.jpg */
                            backdrop_path?: string;
                            genre_ids?: number[];
                            /**
                             * @default 0
                             * @example 550
                             */
                            id: number;
                            /** @example en */
                            original_language?: string;
                            /** @example Fight Club */
                            original_title?: string;
                            /** @example A ticking-time-bomb insomniac and a slippery soap salesman channel primal male aggression into a shocking new form of therapy. Their concept catches on, with underground "fight clubs" forming in every town, until an eccentric gets in the way and ignites an out-of-control spiral toward oblivion. */
                            overview?: string;
                            /**
                             * @default 0
                             * @example 73.433
                             */
                            popularity: number;
                            /** @example /pB8BM7pdSp6B6Ih7QZ4DrQ3PmJK.jpg */
                            poster_path?: string;
                            /** @example 1999-10-15 */
                            release_date?: string;
                            /** @example Fight Club */
                            title?: string;
                            /**
                             * @default true
                             * @example false
                             */
                            video: boolean;
                            /**
                             * @default 0
                             * @example 8.433
                             */
                            vote_average: number;
                            /**
                             * @default 0
                             * @example 26279
                             */
                            vote_count: number;
                        }[];
                        /**
                         * @default 0
                         * @example 2
                         */
                        total_pages: number;
                        /**
                         * @default 0
                         * @example 39
                         */
                        total_results: number;
                    };
                };
            };
        };
    };
// ...
    "search-tv": {
        parameters: {
            query: {
                query: string;
                /** @description Search only the first air date. Valid values are: 1000..9999 */
                first_air_date_year?: number;
                include_adult?: boolean;
                language?: string;
                page?: number;
                /** @description Search the first air date and all episode air dates. Valid values are: 1000..9999 */
                year?: number;
            };
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description 200 */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": {
                        /**
                         * @default 0
                         * @example 1
                         */
                        page: number;
                        results?: {
                            /**
                             * @default true
                             * @example false
                             */
                            adult: boolean;
                            /** @example /bsNm9z2TJfe0WO3RedPGWQ8mG1X.jpg */
                            backdrop_path?: string;
                            genre_ids?: number[];
                            /**
                             * @default 0
                             * @example 1396
                             */
                            id: number;
                            origin_country?: string[];
                            /** @example en */
                            original_language?: string;
                            /** @example Breaking Bad */
                            original_name?: string;
                            /** @example When Walter White, a New Mexico chemistry teacher, is diagnosed with Stage III cancer and given a prognosis of only two years left to live. He becomes filled with a sense of fearlessness and an unrelenting desire to secure his family's financial future at any cost as he enters the dangerous world of drugs and crime. */
                            overview?: string;
                            /**
                             * @default 0
                             * @example 298.884
                             */
                            popularity: number;
                            /** @example /ggFHVNu6YYI5L9pCfOacjizRGt.jpg */
                            poster_path?: string;
                            /** @example 2008-01-20 */
                            first_air_date?: string;
                            /** @example Breaking Bad */
                            name?: string;
                            /**
                             * @default 0
                             * @example 8.879
                             */
                            vote_average: number;
                            /**
                             * @default 0
                             * @example 11536
                             */
                            vote_count: number;
                        }[];
                        /**
                         * @default 0
                         * @example 1
                         */
                        total_pages: number;
                        /**
                         * @default 0
                         * @example 1
                         */
                        total_results: number;
                    };
                };
            };
        };
    };
// ...
}